Create WizardPagination component file under WizardComponents folder
